Update patch set 3
Patch Set 3:

It is questionable if we should pull in all of this even if not used anywhere also since maintaining this will create a large amount of headaches. Better to start iterative and add only whats used sonewhere. Adding this to framework will also create the issue of breaking AOSP compatibility if used in apps. while code reuse sounds nice one of our main goals is to keep changes AOSP (SDK) compatible to allow using it on different roms without needing to pull in a lot of dependencies.

Patch-set: 3