Update patch set 3
Patch Set 3: Code-Review-1

I agree with Max. If there was an implementation to be made then yeah but to add it just to add it, seems wasteful and could only cause problems going forward.

Maybe get something going with OmniGears or another app and put this to use first before trying to get this merge......

Patch-set: 3
Label: Code-Review=-1